requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
openai>=2.0.0
python-dotenv>=1.0.0
langfuse>=3.0.0
httpx>=0.27.0
streamlit>=1.36.0
pytest>=8.0.0
